refactor: `useLoadCart` 커스텀 훅으로 분리
+ 상수화, 개행 수정
